fixup

+++ code/Connectors/CompressedBitConnector/inc/CompressedBitConnector.h
@@ -5,15 +5,15 @@
 
 struct IndexBitCounter
 {
-	UInt32		Last;
-	UInt32		countValue255;
-	UInt32		countInt8;
-	UInt32		countInt16;
-	UInt32		countInt32;
-	UInt32		count254BaseInt8;
-	UInt32		count254BaseInt16;
-	UInt32		count253BaseInt8;
-	UInt32		count253BaseInt16;	
+	UInt64		Last;
+	UInt64		countValue255;
+	UInt64		countInt8;
+	UInt64		countInt16;
+	UInt64		countInt32;
+	UInt64		count254BaseInt8;
+	UInt64		count254BaseInt16;
+	UInt64		count253BaseInt8;
+	UInt64		count253BaseInt16;	
 	UInt32		maxIndex;
 };
 struct CompressedBitConnectorHeader: public GML::ML::CacheHeader
